+++ CHANGELOG.rst
@@ -13,8 +13,10 @@ and this project adheres to
 Fixed
 ^^^^^^^
 
-- Fixed the bug that the "node_func" attribute can be overwritten during editing.
-- Fixed the bug that the "output" attribute of the edge is not updated during editing.
+- Fixed the issue that the "node_func" attribute can be overwritten during editing.
+- Fixed the issue that the "output" attribute of the edge is not updated during editing.
+- Fixed the issue of the different behavior of numpy.ufunc signature due to
+  ``inspect.signature`` change in the Python version 3.11.9.
 
 [0.7.0]
 ------------